LPS-200485 Move ZipReaderFactoryImpl and ZipWriterFactoryImpl to module, and make them as OSGI component instead of spring bean
