+++ modules/apps/portal/portal-zip/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Zip
+Bundle-SymbolicName: com.liferay.portal.zip
+Bundle-Version: 1.0.0
