+++ modules/apps/portal/portal-zip/src/main/java/com/liferay/portal/zip/internal/writer/factory/ZipWriterFactoryImpl.java
@@ -3,16 +3,20 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.zip;
+package com.liferay.portal.zip.internal.writer.factory;
 
 import com.liferay.portal.kernel.zip.ZipWriter;
 import com.liferay.portal.kernel.zip.ZipWriterFactory;
+import com.liferay.portal.zip.ZipWriterImpl;
 
 import java.io.File;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Raymond Augé
  */
+@Component(service = ZipWriterFactory.class)
 public class ZipWriterFactoryImpl implements ZipWriterFactory {
 
 	@Override
